Move functions into gantry setup file
